Don't kill process if it already has terminated

+++ tests/pbi_reports/test_pbi_reports.cs
@@ -104,7 +104,8 @@ namespace MuseumBarberini.Analytics.Tests
         /// </summary>
         public void Stop()
             => _logExceptions(() => {
-                Process.Kill();
+                if (!Process.HasExited)
+                    Process.Kill();
             });
 
         /// <summary>
